# webswing-server/webswing-server-war/src/main/resources/openapi/webswing_api.yaml
openapi: 3.0.2
info:
  title: Webswing REST API
  description: The public REST API provided by Webswing Server
  termsOfService: 'https://webswing.org/licensing/'
  contact:
    name: Webswing Limited
    url: 'https://webswing.org/'
  license:
    name: Code licensed under GNU Affero General Public License version 3
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
  version:  API version
servers:
  - url: 'http://localhost:8080/webswing-server'
    description: Default Server Path
  - url: 'http://localhost:8080/webswing-server/swingset3'
    description: Default Application Path for swingset3
  - url: 'http://localhost:8080/webswing-server/netbeans'
    description: Default Application Path for netbeans
...
paths:
...
  '/rest/info':
    get:
      summary: Get Path Info
      description:  Return detailed information of this Server or Application
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationInfo'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getInfo


  #===========PATHS============
  '/rest/paths':
    get:
      summary: Get Available Applications
      description:  Returns basic information about available applications
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BasicApplicationInfo'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getPaths

  #===========STATS============
  '/rest/stats':
    get:
      summary: Get Aggregated Stats
      description:  Stats are used to show usage graphs (CPU, MEM, Net)
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getStats

  #===========ALL SESSIONS============
  '/rest/sessions':
    get:
      summary: Get All Sessions
      description:  Returns all live and finished server sessions
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sessions'
              example:
                sessions: []
                closedSessions: []
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getSessions

  #===========PERMISSIONS============
  '/rest/permissions':
    get:
      summary: Get User Permissions
      description:  Resolves if current user has permissions to perform selected actions
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permissions'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getPermissions

  #===========VERSION============
  '/rest/adminConsoleUrl':
    get:
      summary: URL of admin console
      description:  URL of admin console from server configuration
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example:
                'http://localhost:9004/admin/#/'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getAdminConsoleUrl

  #===========VERSION============
...
  '/apps':
    get:
      summary: Get Applications
      description:  Returns applications current user has access to. Used for displaying application selector page
      tags:
        - 'Basic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationInfoMsg'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getApps


  #===========MANIFEST============
...
  '/rest/session/{id}':
    parameters:
      - in: path
        name: id
        description:  session id to work with
        required: true
        allowEmptyValue: false
        schema:
          type: string
          nullable: false
    get:
      summary: Get Session details
      description:  Returns details of single session
      tags:
        - 'Manage Sessions'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwingSession'
        '204':
          description: OK
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getSession
    delete:
      summary: Request session shutdown
      description:  Requests clean session shutdown or forced session process kill if force parameter is true
      tags:
        - 'Manage Sessions'
      parameters:
        - in: query
          name: force
          description: Will be parsed as boolean.
          required: false
          allowEmptyValue: true
          schema:
            type: string
            nullable: true
      responses:
        '204':
          description: OK
        '401':
          description: User is not allowed to execute the action in this REST service.
        '500':
          description: Instance with id not found.
      operationId: shutdown


  #===========SESSION METRICS============
...
  '/rest/record/{id}':
    parameters:
      - in: path
        name: id
        description:  Instance Id
        required: true
        allowEmptyValue: false
        schema:
          type: string
          nullable: false
    get:
      summary: Start Session Recording
      description:  Requests server to start session recording
      tags:
        - 'Manage Sessions'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwingSession'
        '204':
          description: OK
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: startRecording


  #===========THREAD DUMP============
  '/rest/threadDump/{instanceId}':
    parameters:
      - in: path
        name: instanceId
        description:  description
        required: true
        allowEmptyValue: false
        schema:
          type: string
          nullable: false
    get:
      summary: Get Thread Dump
      description:  Returns content of existing thread dump specified by timestamp
      tags:
        - 'Manage Sessions'
      parameters:
        - in: query
          name: timestamp
          description:  timestamp when this threaddump was created
          required: false
          allowEmptyValue: false
          schema:
            type: string
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: User is not allowed to execute the action in this REST service.
        '404':
          description: Not found.
      operationId: getThreadDump
    post:
      summary:  Create Thread Dump
      description:  Requests server to create a new thread dump for selected instance ID
      tags:
        - 'Manage Sessions'
      responses:
        '204':
          description: OK
        '401':
          description: User is not allowed to execute the action in this REST service.
        '404':
          description: Not found.
      operationId: requestThreadDump


  #===========CSRF TOKEN============
  '/rest/CSRFToken':
    get:
      summary: CSRF Token
      description:  Returns token required to establish websocket connection
      tags:
        - 'Manage Sessions'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example:
                'Ce5R1E0RSmffU85Jqs4DUpyzERQVLQRgftu5C3qO+NE='
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: generateCsrfToken


  #===========LIST SESSION LOGS============
  '/rest/session/logs/instanceIds':
    get:
      summary: Get session-logging Instances
      description:  Returns instanceIds with per-session logging enabled
      tags:
        - 'Manage Sessions'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getLogInstanceIds

  #===========SESSION LOGS============
  '/rest/session/logs':
    get:
      summary: Download Session Logs
      description: Downloads logs file as a single zip file
      tags:
        - 'Manage Sessions'
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
      operationId: downloadSessionsLog

    post:
      summary:  Request Session Log Content
      description:  Returns requested range of logs from defined session
      tags:
        - 'Manage Sessions'
      requestBody:
        description:  Log Request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
        '401':
          description: User is not allowed to execute the action in this REST service.
        '500':
          description: Failed to read log file.
      operationId: getSessionLogs
  # ============================================================================
  # VIEW LOGS
  # ============================================================================
  #

  #===========LOGS============
  '/rest/logs/{type}':
    parameters:
      - in: path
        name: type
        required: true
        schema:
          type: string
    get:
      summary: Download Logs
      description: Downloads logs file as a single zip file
      tags:
        - 'View Logs'
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
      operationId: downloadLog

    post:
      summary:  Request Log Content
      description:  Returns requested range of logs from defined log type
      tags:
        - 'View Logs'
      requestBody:
        description:  Log Request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
        '401':
          description: User is not allowed to execute the action in this REST service.
        '500':
          description: Failed to read log file.
      operationId: getLogs

  #===========SESSION LOGGING APPS============
  '/rest/logs/sessionApps':
    get:
      summary: Get Session-Logging Applications
      description:  Returns list of applications with per session logging enabled
      tags:
        - 'View Logs'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BasicApplicationInfo'
        '401':
          description: User is not allowed to execute the action in this REST service.
      operationId: getAppsForSessionLogView

# ==============================================================================
# COMPONENTS
# ==============================================================================
#
components:
...
  schemas:
...
    BasicApplicationInfo:
      type: object
      properties:
        path:
          type: string
        url:
          type: string
        name:
          type: string
        runningInstances:
          type: integer
        enabled:
          type: boolean


    # --------------------------------------------------------------------------
    # ApplicationInfo
    # --------------------------------------------------------------------------
    #
    ApplicationInfo:
      type: object
      properties:
        path:
          type: string
        url:
          type: string
        name:
          type: string
        runningInstances:
          type: integer
        enabled:
          type: boolean
        message:
          type: string
        icon:
          type: string
          format: byte
        status:
          $ref: '#/components/schemas/InstanceManagerStatus'
        config:
          type: object
        variables:
          type: object
          additionalProperties:
            type: string
        connectedInstances:
          type: integer
        maxRunningInstances:
          type: integer
        finishedInstances:
          type: integer
        stats:
          type: object
          additionalProperties:
            type: number
            additionalProperties:
              type: number
        warnings:
          type: object
          additionalProperties:
            type: array
            items:
              type: string


    # --------------------------------------------------------------------------
    # ApplicationInfoMsg
    # --------------------------------------------------------------------------
    #
    ApplicationInfoMsg:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        base64Icon:
          type: string
          format: byte


    # --------------------------------------------------------------------------
    # InstanceManagerStatus
    # --------------------------------------------------------------------------
    #
    InstanceManagerStatus:
      type: object
      properties:
        status:
          type: string
          enum:
            - Starting
            - Running
            - Stopped
            - Stopping
            - Error
        error:
          type: string
        errorDetails:
          type: string


    # --------------------------------------------------------------------------
    # LogRequest
    # --------------------------------------------------------------------------
    #
    LogRequest:
      type: object
      properties:
        backwards:
          type: boolean
        offset:
          type: integer
          format: int64
        max:
          type: integer
          format: int64
        instanceId:
          type: string

    # --------------------------------------------------------------------------
    # LogResponse
    # --------------------------------------------------------------------------
    #
    LogResponse:
      type: object
      properties:
        startOffset:
          type: integer
          format: int64
        endOffset:
          type: integer
          format: int64
        log:
          type: string


    # --------------------------------------------------------------------------
    # MetaField
    # --------------------------------------------------------------------------
    #
...
    Sessions:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SwingSession'
        closedSessions:
          type: array
          items:
            $ref: '#/components/schemas/SwingSession'
        recordings:
          type: array
          items:
            type: string

    # --------------------------------------------------------------------------
    # SwingSession
    # --------------------------------------------------------------------------
    #
    SwingSession:
      type: object
      properties:
        serialVersionUID:
          type: integer
          format: int64
          default: 147477596803123012
        id:
          type: string
        user:
          type: string
        userIp:
          type: string
        userOs:
          type: string
        userBrowser:
          type: string
        application:
          type: string
        applicationPath:
          type: string
        startedAt:
          type: integer
          format: int64
        endedAt:
          type: integer
          format: int64
        connected:
          type: boolean
        applet:
          type: boolean
        disconnectedSince:
          type: integer
          format: int64
        recorded:
          type: boolean
        recordingFile:
          type: string
        stats:
          type: string
          additionalProperties: true
        metrics:
          type: object
          additionalProperties:
            type: number
        status:
          type: string
          enum:
            - NOT_STARTED
            - EXITING
            - RUNNING
            - FORCE_KILLED
            - FINISHED
        warnings:
          type: array
          items:
            type: string
        warningHistory:
          type: array
          items:
            type: string
        threadDumps:
          type: object
          additionalProperties:
            type: string
        applicationUrl:
          type: string
        loggingEnabled:
          type: boolean


    # --------------------------------------------------------------------------
    # VariableSetName
    # --------------------------------------------------------------------------
    #
...
    Stats:
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: number


    # --------------------------------------------------------------------------
    # Permissions
    # --------------------------------------------------------------------------
    #
    Permissions:
      description: |
        Permissions map.
        * key - string
        * value - boolean
      type: object
      properties:
        dashboard:
          type: boolean
        configView:
          type: boolean
        configSwingEdit:
          type: boolean
        sessions:
          type: boolean
        configEdit:
          type: boolean
          default: false
        start:
          type: boolean
          default: false
        stop:
          type: boolean
          default: false
        remove:
          type: boolean
          default: false
        create:
          type: boolean
          default: false
        logsView:
          type: boolean
          default: false

    # --------------------------------------------------------------------------
    # Manifest
    # --------------------------------------------------------------------------
    #
...